+++ types.go
@@ -406,7 +406,6 @@ type Checkout struct {
 	Currency    string         `json:"currency"`
 	Chain       ChainId        `json:"chain"`
 	Token       TokenSymbol    `json:"token"`
-	Description *string        `json:"description"`
 	SuccessURL  string         `json:"successUrl"`
 	CancelURL   *string        `json:"cancelUrl"`
 	ExpiresAt   string         `json:"expiresAt"`